Introduce smart reverse and macOS support
Added a new PPM control type: Current Hyst Reverse With Brake.

In this mode the output is off when the input is centered.
Input less than center brakes until the motor stops, at which point it
starts in the reverse direction, but if Max dir switch ERPM is enabled
it will stop the reverse when it reaches the Max ERPM for direction switch.