Added autoscan if CAN list unpopulated
Also changed name to motor ID. Think this will be more clear to new users that they are selecting different motors.  Also throw a warning if scan can clicked when not connected and clear the CAN list.

Also I added the other DRVs to reset latched faults... kind of a hack to send them all but gets the job done.