Configurable input voltage correction factor
The resistors used to measure input voltage are often not very accurate
causing noticeable errors in the reported input voltage.

This feature lets users specify a correction factor, expressed as a
percentage. Now the input voltage used everywhere is Vin * (1 + corr).

The correction factor is limited to values in the range of -10% .. +10%

Signed-off-by: Dado Mista <dadomista@gmail.com>