Package loading fixes (#348)
* Check package was loaded ok when opening the file

Without the check, a zero-size file would open without error, allowing
to write it as well.

* Make CodeLoader::installVescPackage() always erase both qml and lisp

If one or the other is missing in the package, we still want to erase
the memory, otherwise the code of the old package will stay there and
keep running.

If a package that contains neither qml nor lisp is loaded, it erases
both. That should be the expected outcome.

Reorganizes the logic a bit to hopefully make it more streamlined.