Refactor flake overlay into pkgs/default.nix
Moves the definition of the packages attribute set into
"pkgs/default.nix", making the overlay import it. "flake.nix" now also
imports it directly instead of going through the overlay. I don't really
know, but this may improve the evaluation performance slightly. :)
But it mainly makes is so that the package set definition is only in
one file.