Update flake to nixpkgs 25.05
Updates the flake nixpkgs input to 25.05, and fixes the errors caused by
it.

Unfortunately gcc-arm-embedded-7, used when building the bldc firmwares,
has been removed from nixpkgs, so we now also include an older version
of it for gcc-arm-embedded-7 specifically.
Simply trying to build bldc with the newer gcc version 14 breaks for
some reason, with the linker giving an error about overlaping sections
that I'm not knowledgeable enough to solve. And since the project does
use that version specifically keeping the old version feels like the
correct solution for now.

Going forwards I think the best solution would be to migrate the project
to use a version of gcc-arm-embedded >= 13, but that's of course easier
said than done.