Add support for .lbm extension
Adds support for the ".lbm" file extension, prefering it over the old
".lisp".

Concretely, this commit adds support for ".lbm":
- in all file dialogs
- when creating lbm and vesc packages
- when reducing lbm files
- when packing LispBM imports.

The LispBM file info dialog's text is updated to reflect this.

It also adds preference for the ".lbm" extension in the "From VESC.lbm"
filename and in some debug files' names.