Change usage of "Lisp" to "LispBM" in UI
Changes all places which refer to LispBM as "Lisp" to instead use
the full capitalized form "LispBM" to be more precise. This aligns the
UI with the change to prefer the ".lbm" extension over ".lisp".

Also removes incorrectly hyphens in compound nouns which include "lisp".
For example, "LispBM-scripts" is changed to "LispBM scripts".

Additionally changes the Discord channel name referenced in pagelisp.cpp to
match the name in the official Discord.