Stable sorting and icons for connect screen
This addresses items jumping up and down in the connect screen when
devices appear/disappear for example on LAN. It also adds different
icons depending on the connection type. To avoid flickering, the model
is only updated if it is changed.

I also took the opportunity to fix the binding loop errors in this qml,
as well as cleaning up some of the javascript code for model building.